requirements.txt
django
psycopg[binary]
python-dotenv
icalendar
python-dateutil
djangorestframework
pillow
sphinx